requirements.txt
gym==0.26.2
matplotlib
numpy
pyyaml
tensorboardX
torch==1.9.1
wandb>=0.17.4
